Back to using the high-resolution SDL timer instead of the previous low-resolution one. It turned out that the low-resolution one was sometimes returning the same time as the previous tick, which lead to some floats becoming nan (not a number), which unsigned chars (bytes) couldn't represent (that was the bug).

+++ src/audio_visual/accel/src/headers/camera.h
@@ -8,7 +8,7 @@
 //////////
 
 typedef struct {
-	Uint32 last_time;
+	Uint64 last_time;
 
 	struct {
 		GLfloat fov, hori, vert, tilt;
